refactor: reorganize dashboard views
